+++ src/simply/simply.h
@@ -7,6 +7,7 @@ typedef struct Simply Simply;
 struct Simply {
   struct SimplyAccel *accel;
   struct SimplyTouch *touch;
+  struct SimplyAudio *audio;
   struct SimplyVoice *voice;
   struct SimplyRes *res;
   struct SimplyMsg *msg;