+++ src/simply/simply_msg.c
@@ -2,6 +2,7 @@
 
 #include "simply_accel.h"
 #include "simply_touch.h"
+#include "simply_audio.h"
 #include "simply_voice.h"
 #include "simply_res.h"
 #include "simply_stage.h"
@@ -315,6 +316,7 @@ static void handle_packet(Simply *simply, Packet *packet) {
   if (simply_ui_handle_packet(simply, packet)) { return; }
   if (simply_accel_handle_packet(simply, packet)) { return; }
   if (simply_touch_handle_packet(simply, packet)) { return; }
+  if (simply_audio_handle_packet(simply, packet)) { return; }
   if (simply_voice_handle_packet(simply, packet)) { return; }
   if (simply_menu_handle_packet(simply, packet)) { return; }
   if (simply_stage_handle_packet(simply, packet)) { return; }